feat: restructure views and implement AppLogin, AppHome, AppRegister, and AppCalendar components
